WIP
There is still a limitation with this new approach when it comes to sub
queries. We resolve the location of the where clause keys (docKeys) on
the handler node before all symbols which might be contained in the
docKeys are resolved. That's why we are still issue GetRequests for
these keys. These are performed on the handler node. This extra step can
be removed once the runtime supports deferred plans which allow to build
up a plan depending on its dependencies.