Add a jobs_log filter setting to control entries being recorded
Depending on the retention policy (size or expiration), `sys.jobs_log`
might be flooded or grow too large in size on a high traffic cluster.

This introduces a new `stats.jobs_log_filter` setting which can contain
an expression evaluated against a `sys.jobs_log` record to determine if
it should be included in `sys.jobs_log`.

This can for example be used to only include entries with a certain
duration (`ended - started > 30`)